requirements.txt
# fastapi[standard]==0.115.0
# uvicorn[standard]==0.30.1
# pydantic==2.8.2
# pydantic-settings==2.1.0
# pytest==8.2.2
# pytest-asyncio==0.21.1
# httpx==0.27.0
# psycopg2==2.9.11
# requests==2.31.0
# testcontainers==3.7.0
# sqlalchemy==2.0.25

fastapi[standard]==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.1.0
pytest==8.2.2
pytest-asyncio==0.21.1
httpx==0.27.0
psycopg2-binary==2.9.11
requests==2.31.0
testcontainers==3.7.0
sqlalchemy==2.0.25
alembic==1.17.2
python-jose[cryptography]==3.3.0
cryptography==42.0.0